+++ geostore_routing/functions.py
@@ -1,0 +1,5 @@
+from django.contrib.gis.db.models.functions import GeoFunc
+
+
+class LineSubstring(GeoFunc):
+    pass